requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-MySQLdb==0.2.0
gunicorn==20.1.0
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.25
mysqlclient==2.0.3
protobuf==3.17.3
six==1.16.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
